Fixed #6: long-addressed payload segments. Upgraded database format.

+++ core/src/main/java/com/yandex/yoctodb/util/OutputStreamWritable.java
@@ -23,7 +23,7 @@ import java.io.OutputStream;
  */
 @ThreadSafe
 public interface OutputStreamWritable {
-    int getSizeInBytes();
+    long getSizeInBytes();
 
     void writeTo(
             @NotNull
